fix(dsv4 fp8 prefill): bf16-round normed rope tail before rotate in fused-kv pack
The fused-kv producer mirrored the compressed-K store kernel, which rotates the
rope tail in fp32. The JIT fused_norm_rope_inplace (the extend path's reference)
writes the normed tail as bf16 and rotates that rounded value, like the q
producer. Rotating in fp32 diverged by ~1 bf16 ULP, amplified by the kv norm
weight, pushing the reconstruction rel error just over 1e-3. Round to bf16
before rotating to match the JIT. NoPE was already byte-exact.

+++ python/sglang/srt/layers/attention/dsv4/unified_kv_kernels/runtime.py
@@ -537,8 +537,12 @@ def _fused_kv_norm_rope_mxfp8_pack_kernel(
     w_imag = tl.load(
         weight_ptr + imag_cols, mask=pair_offs < (DIM_ROPE // 2), other=1.0
     ).to(tl.float32)
-    x_real = x_real * rms_inv[:, None] * w_real[None, :]
-    x_imag = x_imag * rms_inv[:, None] * w_imag[None, :]
+    # Round the normed (norm*weight) rope value to bf16 BEFORE rotating: the JIT
+    # ``fused_norm_rope_inplace`` writes the normed tail as bf16 and rotates that
+    # rounded value (mirrors the q producer). Rotating in fp32 here diverges by
+    # ~1 bf16 ULP, amplified by the (randn-scale) weight. Match the JIT instead.
+    x_real = (x_real * rms_inv[:, None] * w_real[None, :]).to(tl.bfloat16).to(tl.float32)
+    x_imag = (x_imag * rms_inv[:, None] * w_imag[None, :]).to(tl.bfloat16).to(tl.float32)
 
     positions = tl.load(positions_ptr + rows, mask=rmask, other=0).to(tl.int64)
     freq_base = positions[:, None] * freq_row_stride
